[tree] Move RFriendInfo out of the Internal namespace
As in https://github.com/root-project/root/pull/11465 we would like to
have a public getter returning RFriendInfo object. Hence, we move the
class out of the Internal namespace (in its own header file).
