+++ tree/tree/inc/ROOT/InternalTreeUtils.hxx
@@ -21,6 +21,7 @@
 #include "TChain.h"
 #include "TNotifyLink.h"
 #include "TObjArray.h"
+#include "ROOT/RFriendInfo.hxx"
 
 #include <memory>
 #include <string>
@@ -38,41 +39,9 @@ namespace Internal {
 */
 namespace TreeUtils {
 
-using NameAlias = std::pair<std::string, std::string>; ///< A pair of name and alias of a TTree's friend tree.
-/**
-\struct ROOT::Internal::TreeUtils::RFriendInfo
-\brief Information about friend trees of a certain TTree or TChain object.
-\ingroup tree
-*/
-struct RFriendInfo {
-
-   std::vector<NameAlias> fFriendNames; ///< Pairs of names and aliases of friend trees/chains.
-   /**
-   Names of the files where each friend is stored. fFriendFileNames[i] is the
-   list of files for friend with name fFriendNames[i].
-   */
-   std::vector<std::vector<std::string>> fFriendFileNames;
-   /**
-      Names of the subtrees of a friend TChain. fFriendChainSubNames[i] is the
-      list of names of the trees that make a friend TChain whose information is
-      stored at fFriendNames[i] and fFriendFileNames[i]. If instead the friend
-      tree at position `i` is a TTree, fFriendChainSubNames[i] will be an empty
-      vector.
-   */
-   std::vector<std::vector<std::string>> fFriendChainSubNames;
-
-   void AddFriend(const std::string &treeName, const std::string &fileNameGlob, const std::string &alias = "");
-
-   void
-   AddFriend(const std::string &treeName, const std::vector<std::string> &fileNameGlobs, const std::string &alias = "");
-
-   void AddFriend(const std::vector<std::pair<std::string, std::string>> &treeAndFileNameGlobs,
-                  const std::string &alias = "");
-};
-
 std::vector<std::string> GetTopLevelBranchNames(TTree &t);
 std::vector<std::string> GetFileNamesFromTree(const TTree &tree);
-RFriendInfo GetFriendInfo(const TTree &tree);
+ROOT::TreeUtils::RFriendInfo GetFriendInfo(const TTree &tree);
 std::vector<std::string> GetTreeFullPaths(const TTree &tree);
 
 void ClearMustCleanupBits(TObjArray &arr);
